+++ src/extension.ts
@@ -1,6 +1,8 @@
 import * as vscode from 'vscode'; // VS Code Extension API
 import { comparePermissions } from './command/comparePermissions'; // Import sub-module
-import { openScratchOrg } from './command/openScratchOrg';
+import { validateChanges } from './command/validateChanges'; // function
+import { listAllOrgs } from './command/listOrgs'; // function
+import { OrgType, Operation } from './command/listOrgs'; // Enum
 // import { promisify } from 'util'; // Allows to use note-util's promise function with async/await to get shell commands which read immediately.
 
 /**
@@ -15,7 +17,15 @@ export function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(vscode.commands.registerCommand('msk.comparePermissions', comparePermissions));
 
 	/* Opens a Scratch Org or Dev Hub without setting it as default. */
-	context.subscriptions.push(vscode.commands.registerCommand('msk.openScratchOrg', openScratchOrg));
+	context.subscriptions.push(vscode.commands.registerCommand('msk.openScratchOrg', function execute() { listAllOrgs(OrgType.ScratchDev, Operation.Open); }));
+	// context.subscriptions.push(vscode.commands.registerCommand('msk.openScratchOrg', openScratchOrg));
+
+	/* Deletes a Scratch Org. */
+	context.subscriptions.push(vscode.commands.registerCommand('msk.deleteScratchOrg', function execute() { listAllOrgs(OrgType.Scratch, Operation.Delete); }));
+	// context.subscriptions.push(vscode.commands.registerCommand('msk.deleteScratchOrg', validateChanges));
+
+	/* Validates local changes against Scratch Org and Sandbox. */
+	context.subscriptions.push(vscode.commands.registerCommand('msk.validateChanges', validateChanges));
 }
 
 /**